# perf!: optimize mesgdef structs