--- a/profile/mesgdef/chrono_shot_data_gen.go
+++ b/profile/mesgdef/chrono_shot_data_gen.go
@@ -9,7 +9,6 @@ package mesgdef
 import (
 	"github.com/muktihari/fit/factory"
 	"github.com/muktihari/fit/kit/datetime"
-	"github.com/muktihari/fit/kit/scaleoffset"
 	"github.com/muktihari/fit/profile/basetype"
 	"github.com/muktihari/fit/profile/typedef"
 	"github.com/muktihari/fit/proto"
@@ -107,7 +106,7 @@ func (m *ChronoShotData) ShotSpeedScaled() float64 {
 	if m.ShotSpeed == basetype.Uint32Invalid {
 		return math.Float64frombits(basetype.Float64Invalid)
 	}
-	return scaleoffset.Apply(m.ShotSpeed, 1000, 0)
+	return float64(m.ShotSpeed)/1000 - 0
 }
 
 // SetTimestamp sets Timestamp value.
@@ -129,7 +128,12 @@ func (m *ChronoShotData) SetShotSpeed(v uint32) *ChronoShotData {
 //
 // Scale: 1000; Units: m/s
 func (m *ChronoShotData) SetShotSpeedScaled(v float64) *ChronoShotData {
-	m.ShotSpeed = uint32(scaleoffset.Discard(v, 1000, 0))
+	unscaled := (v + 0) * 1000
+	if math.IsNaN(unscaled) || math.IsInf(unscaled, 0) || unscaled > float64(basetype.Uint32Invalid) {
+		m.ShotSpeed = uint32(basetype.Uint32Invalid)
+		return m
+	}
+	m.ShotSpeed = uint32(unscaled)
 	return m
 }