--- a/profile/mesgdef/device_aux_battery_info_gen.go
+++ b/profile/mesgdef/device_aux_battery_info_gen.go
@@ -9,7 +9,6 @@ package mesgdef
 import (
 	"github.com/muktihari/fit/factory"
 	"github.com/muktihari/fit/kit/datetime"
-	"github.com/muktihari/fit/kit/scaleoffset"
 	"github.com/muktihari/fit/profile/basetype"
 	"github.com/muktihari/fit/profile/typedef"
 	"github.com/muktihari/fit/proto"
@@ -121,7 +120,7 @@ func (m *DeviceAuxBatteryInfo) BatteryVoltageScaled() float64 {
 	if m.BatteryVoltage == basetype.Uint16Invalid {
 		return math.Float64frombits(basetype.Float64Invalid)
 	}
-	return scaleoffset.Apply(m.BatteryVoltage, 256, 0)
+	return float64(m.BatteryVoltage)/256 - 0
 }
 
 // SetTimestamp sets Timestamp value.
@@ -149,7 +148,12 @@ func (m *DeviceAuxBatteryInfo) SetBatteryVoltage(v uint16) *DeviceAuxBatteryInfo
 //
 // Scale: 256; Units: V
 func (m *DeviceAuxBatteryInfo) SetBatteryVoltageScaled(v float64) *DeviceAuxBatteryInfo {
-	m.BatteryVoltage = uint16(scaleoffset.Discard(v, 256, 0))
+	unscaled := (v + 0) * 256
+	if math.IsNaN(unscaled) || math.IsInf(unscaled, 0) || unscaled > float64(basetype.Uint16Invalid) {
+		m.BatteryVoltage = uint16(basetype.Uint16Invalid)
+		return m
+	}
+	m.BatteryVoltage = uint16(unscaled)
 	return m
 }