### Background 
The objective of this PR is to improve performance by reducing the size of mesgdef's structs through these following approaches:
1. Optimizing the field we use for tracking expanded fields. 
Previously, we use array of bool, e.g. `[109]bool` for Record. We found this is not efficient since 1 bool is equal to 1 byte, meanwhile 1 byte has 8 bits that we can use for flag, wasting 7 bits for every single bool, this can be compacted by using `[14]byte` that can hold up to 14*8=112 flags (the size of array is vary depends on each struct's requirement).
2. Use array instead of slice for fixed array since most of fixed arrays are small-sized with the type that only consume small memory, e.g. Record's CompressedSpeedDistance is [3]byte which equal to 3 bytes. Previously, we use slice that requires 24 bytes (8 for pointer to array, 8 for len, 8 for cap), we save 21 bytes for this case. Moreover we ensure that users' input is valid, e.g. CompressedSpeedDistance must contains exactly 3 elements or no element at all so it can be omitted.

### Result
Most structs' size are reduced but few of them are being increased instead. However, we ensure the hot path structs are being optimized, such as Record. (Note: A hot path struct is a struct that being instantiated many times or frequently used)

The size of Record struct was 464 bytes, now it is 352 bytes (contains 5 bytes padding, means we can still add fields that worth <= 5 bytes). For size of 464 bytes, we need class 25 which is 480 bytes while 352 bytes can fit perfectly to class 21 which is 352 bytes, reducing 128 bytes per instance. Please see the following reference (https://go.dev/src/runtime/sizeclasses.go): 
```
// class  bytes/obj  bytes/span  objects  tail waste  max waste  min align
...
//    21        352        8192       23          96      9.88%         32
//    22        384        8192       21         128      9.51%        128
//    23        416        8192       19         288     10.71%         32
//    24        448        8192       18         128      8.37%         64
//    25        480        8192       17          32      6.82%         32
...
```

By reducing the size of the struct that can fit to smaller class, we are reducing runtime's burden to allocate memory.

### Benchmark
```js
goos: darwin; goarch: amd64; pkg: benchfit
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                       │   old.txt   │              new.txt               │
                       │   sec/op    │   sec/op     vs base               │
Decode/muktihari/fit-4   84.98m ± 3%   82.65m ± 3%  -2.73% (p=0.015 n=10)

                       │   old.txt    │               new.txt                │
                       │     B/op     │     B/op      vs base                │
Decode/muktihari/fit-4   49.78Mi ± 0%   37.57Mi ± 0%  -24.52% (p=0.000 n=10)

                       │   old.txt   │              new.txt               │
                       │  allocs/op  │  allocs/op   vs base               │
Decode/muktihari/fit-4   100.2k ± 0%   100.2k ± 0%       ~ (p=1.000 n=10)

```

### Additional changes

Additional change that has to be made due to second approach of the PR's goal is generating scale and offset calculation directly to all _Scaled_ method and add missing input validation before the calculation process for slice/array. The _scaleoffset_ package's functions do not do it for us and it will be tricky to do it in the _scaleoffset_ package without affecting the performance with the current Go's generic implementation.

### Breaking Changes 
1. mesgdef: Field `IsExpandedFields [N]bool` is replaced by `IsExpanded(fieldNum byte) bool` method to get the value and `MarkAsExpandedField(fieldNum byte, flag bool) (ok bool)` method to set the value (hiding the implementation details from users, hopefuly we don't break users' code next time we need to change the logic).
2. mesgdef: Fixed array is now represented as fixed array (previously was represented using slice), e.g. Record's CompressedSpeedDistance is now [3]byte instead of []byte.